http://issues.carrot2.org/browse/CARROT-657
Separate boolean[] array for marking stop words removed. Integrated query flag with token flags. AllWords.type is now short[].



git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@4225 7ff1d41c-760d-0410-a7ff-a3a56f310b35

+++ core/carrot2-util-text/src-test/org/carrot2/text/preprocessing/CaseNormalizerTest.java
@@ -208,7 +208,7 @@ public class CaseNormalizerTest extends PreprocessingComponentTestBase
                 0
             }
         };
-        int [] expectedType = new int []
+        short [] expectedType = new short []
         {
             ITokenTypeAttribute.TT_NUMERIC, ITokenTypeAttribute.TT_TERM, ITokenTypeAttribute.TT_FILE,
             ITokenTypeAttribute.TT_BARE_URL, ITokenTypeAttribute.TT_EMAIL
@@ -703,7 +703,7 @@ public class CaseNormalizerTest extends PreprocessingComponentTestBase
 
     private void check(char [][] expectedWordImages, int [] expectedWordTf,
         int [] expectedWordIndices, int [][] expectedWordTfByDocument,
-        byte [][] expectedFieldIndex, int [] expectedType)
+        byte [][] expectedFieldIndex, short [] expectedType)
     {
         tokenizer.tokenize(context);
         caseNormalizer.normalize(context);
@@ -727,10 +727,10 @@ public class CaseNormalizerTest extends PreprocessingComponentTestBase
         temporaryCaseNormalizer.dfThreshold = caseNormalizer.dfThreshold;
     }
 
-    private int [] createTermTokenTypes(int count)
+    private short [] createTermTokenTypes(int count)
     {
-        final int [] result = new int [count];
-        Arrays.fill(result, ITokenTypeAttribute.TT_TERM);
+        final short [] result = new short [count];
+        Arrays.fill(result, (short) ITokenTypeAttribute.TT_TERM);
         return result;
     }
 }
